refactor: separate page component
